core: Split of CellStructure and made it self-contained
