+++ src/core/collision.cpp
@@ -663,7 +663,7 @@ void handle_collisions() {
 
     // If any node had a collision, all nodes need to resort
     if (!gathered_queue.empty()) {
-      set_resort_particles(Cells::RESORT_GLOBAL);
+      cell_structure.set_resort_particles(Cells::RESORT_GLOBAL);
       cells_update_ghosts(GHOSTTRANS_PROPRTS | GHOSTTRANS_BONDS);
     }
   }    // are we in one of the vs_based methods
