+++ src/core/integrators/velocity_verlet_inline.hpp
@@ -56,7 +56,7 @@ inline void velocity_verlet_propagate_vel_pos(const ParticleRange &particles) {
             Utils::sqr(p.r.p[1] - p.l.p_old[1]) +
             Utils::sqr(p.r.p[2] - p.l.p_old[2]) >
         skin2)
-      set_resort_particles(Cells::RESORT_LOCAL);
+      cell_structure.set_resort_particles(Cells::RESORT_LOCAL);
   }
 }
