+++ src/core/short_range_loop.hpp
@@ -97,7 +97,7 @@ void short_range_loop(ParticleKernel &&particle_kernel,
                       const VerletCriterion &verlet_criterion = {}) {
   ESPRESSO_PROFILER_CXX_MARK_FUNCTION;
 
-  assert(get_resort_particles() == Cells::RESORT_NONE);
+  assert(cell_structure.get_resort_particles() == Cells::RESORT_NONE);
 
   if (cell_structure.min_range != INACTIVE_CUTOFF) {
     auto first =
